Extract TypeScript transform into separate config
