+++ packages/babel-preset-expo/src/__tests__/preset-config.test.ts
@@ -58,6 +58,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-inline-or-reference-env-vars",
@@ -81,9 +82,8 @@ describe('plugin list snapshots', () => {
         "transform-private-methods",
         "transform-private-property-in-object",
         "transform-unicode-regex",
-        "transform-typescript",
         "warn-on-deep-imports",
-        "base$0$1.overrides[4]$0",
+        "base$0$1.overrides[2]$0",
         "transform-classes",
         "transform-named-capturing-groups-regex",
         "transform-destructuring",
@@ -115,6 +115,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-class-static-block",
         "expo-define-globals",
         "expo-inline-or-reference-env-vars",
@@ -137,9 +138,8 @@ describe('plugin list snapshots', () => {
         "transform-private-methods",
         "transform-private-property-in-object",
         "transform-unicode-regex",
-        "transform-typescript",
         "warn-on-deep-imports",
-        "base$0$1.overrides[4]$0",
+        "base$0$1.overrides[2]$0",
         "transform-named-capturing-groups-regex",
         "transform-optional-catch-binding",
         "transform-destructuring",
@@ -173,6 +173,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-class-static-block",
         "expo-define-globals",
         "expo-minify-platform-select",
@@ -196,8 +197,7 @@ describe('plugin list snapshots', () => {
         "transform-private-methods",
         "transform-private-property-in-object",
         "transform-unicode-regex",
-        "transform-typescript",
-        "base$0$1.overrides[4]$0",
+        "base$0$1.overrides[2]$0",
         "transform-named-capturing-groups-regex",
         "transform-destructuring",
         "transform-async-generator-functions",
@@ -222,6 +222,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-inline-or-reference-env-vars",
@@ -246,7 +247,6 @@ describe('plugin list snapshots', () => {
         "transform-private-property-in-object",
         "transform-export-namespace-from",
         "transform-runtime",
-        "transform-typescript",
         "proposal-export-default-from",
         "transform-modules-commonjs",
       ]
@@ -264,6 +264,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-minify-platform-select",
@@ -289,7 +290,6 @@ describe('plugin list snapshots', () => {
         "transform-private-property-in-object",
         "transform-export-namespace-from",
         "transform-runtime",
-        "transform-typescript",
         "proposal-export-default-from",
         "transform-modules-commonjs",
       ]
@@ -312,6 +312,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-inline-manifest-plugin",
@@ -334,7 +335,6 @@ describe('plugin list snapshots', () => {
         "transform-private-property-in-object",
         "transform-export-namespace-from",
         "transform-runtime",
-        "transform-typescript",
         "proposal-export-default-from",
         "transform-modules-commonjs",
       ]
@@ -357,6 +357,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-inline-or-reference-env-vars",
@@ -381,7 +382,6 @@ describe('plugin list snapshots', () => {
         "transform-private-property-in-object",
         "transform-export-namespace-from",
         "transform-runtime",
-        "transform-typescript",
         "proposal-export-default-from",
         "transform-modules-commonjs",
       ]
@@ -404,6 +404,7 @@ describe('plugin list snapshots', () => {
         "syntax-optional-chaining",
         "transform-flow-enums",
         "transform-flow-strip-types",
+        "transform-typescript",
         "transform-object-rest-spread",
         "expo-define-globals",
         "expo-inline-or-reference-env-vars",
@@ -428,9 +429,8 @@ describe('plugin list snapshots', () => {
         "transform-private-methods",
         "transform-private-property-in-object",
         "transform-unicode-regex",
-        "transform-typescript",
         "warn-on-deep-imports",
-        "base$0$1.overrides[4]$0",
+        "base$0$1.overrides[2]$0",
         "transform-classes",
         "transform-named-capturing-groups-regex",
         "transform-destructuring",
